+++ cucumber-core/src/main/java/io/cucumber/core/feature/FeatureWithLines.java
@@ -17,7 +17,7 @@ import java.util.stream.Collectors;
  * Identifies either a directory containing feature files, a specific feature or
  * specific scenarios and examples (pickles) in a feature.
  * <p>
- * The syntax of a a feature with lines defined as either a {@link FeaturePath}
+ * The syntax of a feature with lines defined as either a {@link FeaturePath}
  * or a {@link FeatureIdentifier} followed by a sequence of line numbers each
  * preceded by a colon.
  */